[#205] - 강퇴하기 시 pay_status가 2미만인 유저만 강퇴가능, - 채팅방 상태 모델 및 로직에서 제거

+++ gatgu/chat/views.py
@@ -107,7 +107,7 @@ class OrderChatViewSet(viewsets.GenericViewSet):
             if chatting.article.article_status != Article.GATHERING:
                 return Response({"Could not participate"}, status=status.HTTP_404_NOT_FOUND)
                 
-            elif chatting.order_status == 1:
+            elif chatting.article.article_status == Article.GATHERING:
                 ParticipantProfile.objects.create(order_chat=chatting, participant=user, wish_price=wish_price)
                 return Response(status=status.HTTP_201_CREATED)
             else:
@@ -162,6 +162,8 @@ class OrderChatViewSet(viewsets.GenericViewSet):
 
             try:
                 participant = ParticipantProfile.objects.get(order_chat=chatting, participant_id=exile_id)
+                if participant.pay_status >= 2:
+                    return Response('입금대기중인 유저만 강퇴가능합니다.', status=status.HTTP_400_BAD_REQUEST)
                 participant.delete()
                 return Response(status=status.HTTP_200_OK)
             except ParticipantProfile.DoesNotExist:
@@ -207,10 +209,10 @@ class OrderChatViewSet(viewsets.GenericViewSet):
         chat = get_object_or_404(OrderChat, pk=pk)
         if user != chat.article.writer:
             return Response(status=status.HTTP_403_FORBIDDEN)
-        if 'order_status' in data:
-            if data['order_status'] < chat.order_status:
-                print("order status should grow up")
-                return Response(status=status.HTTP_403_FORBIDDEN)
+        # if 'order_status' in data:
+        #     if data['order_status'] < chat.order_status:
+        #         print("order status should grow up")
+        #         return Response(status=status.HTTP_403_FORBIDDEN)
         serializer = self.get_serializer(chat, data=data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.update(chat, serializer.validated_data)
@@ -226,20 +228,20 @@ class OrderChatViewSet(viewsets.GenericViewSet):
         if chatting.article.writer != request.user:
             return Response(status=status.HTTP_403_FORBIDDEN)
 
-        order_status = request.data.get('order_status')
+        # order_status = request.data.get('order_status')
         tracking_number = request.data.get('tracking_number')
 
-        # could not update (both at once or nothing)
-        if (order_status is not None and tracking_number is not None) or (
-                order_status is None and tracking_number is None):
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+        # # could not update (both at once or nothing)
+        # if (order_status is not None and tracking_number is not None) or (
+        #         order_status is None and tracking_number is None):
+        #     return Response(status=status.HTTP_400_BAD_REQUEST)
 
-        if order_status is not None:
-            # validate (order_status's range 1 ~ 4)
-            if order_status not in [i for i, s in OrderChat.ORDER_STATUS]:
-                return Response(status=status.HTTP_400_BAD_REQUEST)
-            else:
-                chatting.order_status = order_status
+        # if order_status is not None:
+        #     # validate (order_status's range 1 ~ 4)
+        #     if order_status not in [i for i, s in OrderChat.ORDER_STATUS]:
+        #         return Response(status=status.HTTP_400_BAD_REQUEST)
+        #     else:
+        #         chatting.order_status = order_status
 
         elif tracking_number is not None:
             chatting.tracking_number = tracking_number
